requirements.txt
opencv-python
pytesseract
pyspellchecker
reportlab
flask
-e .